# docs/swagger.yaml
basePath: /
definitions:
...
  views.Swagger_Comment_Register_Put:
    properties:
      message:
        example: String
        type: string
    type: object
  views.Swagger_Photo_Register_Post:
    properties:
      caption:
        example: String
        type: string
      photo_url:
        example: String
        type: string
      title:
        example: String
        type: string
    type: object
...
host: localhost:os.Getenv("PORT")
info:
  contact:
    email: hansparson013@gmail.com
    name: Hans Parson
  description: Final Project Golang
  termsOfService: http://swagger.io/terms/
  title: My Grams APP
  version: v1.0
paths:
...
  /photos:
    get:
      description: Get Photo
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "201":
          description: success
          schema:
            type: string
      summary: Get Photo
      tags:
      - PHOTO
    post:
      consumes:
      - application/json
      description: Post Photo
      parameters:
      - description: Post Photo
        in: body
        name: Photo
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Photo_Register_Post'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Post Photo
      tags:
      - PHOTO
  /photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete Photo
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id Photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Delete Photo
      tags:
      - PHOTO
    put:
      consumes:
      - application/json
      description: Update Photo
      parameters:
      - description: Update Photo
        in: body
        name: Photo
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Comment_Register_Put'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id Photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Update Photo
      tags:
      - PHOTO
...
swagger: "2.0"